refatorar código e organizar em funções especializadas
