+++ certificado.html
@@ -36,6 +36,6 @@
       feather.replace();
     </script>
 
-    <script src="./scripts/auto-import-components.js"></script>
+    <script type="module" src="./scripts/auto-import-components.js"></script>
   </body>
 </html>
