+++ index.html
@@ -151,7 +151,7 @@
       feather.replace();
   </script>
 
-<script src="./scripts/auto-import-components.js"></script>
+<script type="module" src="./scripts/auto-import-components.js"></script>
 
   </body>
 </html>
