+++ modulo1.html
@@ -95,7 +95,7 @@
     feather.replace();
 </script>
 
-<script src="./scripts/auto-import-components.js" data-module="quiz1.html"></script>
+<script type="module" src="./scripts/auto-import-components.js" data-module="quiz1.html"></script>
 
 </body>
 </html>
