+++ modulo2.html
@@ -308,9 +308,6 @@
 
     <div id="componente-footer"></div>
     
-    <script src="./scripts/auto-import-components.js" data-module="quiz2.html"></script>
-
-    
-    
+    <script type="module" src="./scripts/auto-import-components.js" data-module="quiz2.html"></script>    
 </body>
 </html>
