+++ modulo3.html
@@ -461,7 +461,7 @@
     <div id="componente-footer"></div>
 
     <script src="./scripts/modulo3.js"></script>
-    <script src="./scripts/auto-import-components.js" data-module="quiz3.html"></script>
+    <script type="module" src="./scripts/auto-import-components.js" data-module="quiz3.html"></script>
 </body>
 
 </html>
